# service/openapi.yaml
openapi: 3.1.0
info:
  title: Zailom WA Service
  version: 1.0.0
  summary: Central WhatsApp / Evolution API service — single source of truth for Zailom products.
  description: |
    All endpoints below (except `/healthz`, `/readyz`, `/v1/hooks/*`, and `/v1/admin/*`) require
    a tenant API key in the `Authorization: Bearer <key>` header.

    Admin endpoints require `X-Admin-Token: <ADMIN_TOKEN>`.

    Inbound Evolution webhooks (`/v1/hooks/evolution/:evoName`) are public but
    protected by an unguessable path and optionally by `X-Evolution-Token`.
  contact: { name: Zailom Platform }
servers:
  - url: https://wa.zailom.com
    description: Production
  - url: http://localhost:8080
    description: Local dev
...
security:
  - TenantKey: []
components:
  securitySchemes:
    TenantKey:
      type: http
      scheme: bearer
      bearerFormat: "zwa_live_<prefix>_<secret>"
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
  schemas:
...
    Tenant:
      type: object
      properties:
        id:                { type: string, format: uuid }
        product:           { type: string, enum: [booking, flow, other] }
        product_tenant_id: { type: string }
        name:              { type: string }
        created_at:        { type: string, format: date-time }
...
paths:
...
  /v1/admin/tenants:
    get:
      tags: [Admin]
      security: [{ AdminToken: [] }]
      summary: List tenants
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/Tenant" } } } } } }
    post:
      tags: [Admin]
      security: [{ AdminToken: [] }]
      summary: Create or upsert tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [product, product_tenant_id, name]
              properties:
                product:           { type: string, enum: [booking, flow, other] }
                product_tenant_id: { type: string, example: "company_abc123" }
                name:              { type: string }
                metadata:          { type: object }
            example:
              product: booking
              product_tenant_id: "company_abc123"
              name: "Salão da Ana"
      responses:
        "201":
          description: Created
          content: { application/json: { schema: { $ref: "#/components/schemas/Tenant" } } }
  /v1/admin/api-keys:
    get:
      tags: [Admin]
      security: [{ AdminToken: [] }]
      summary: List API keys
      parameters:
        - in: query
          name: tenant_id
          schema: { type: string, format: uuid }
      responses: { "200": { description: OK } }
    post:
      tags: [Admin]
      security: [{ AdminToken: [] }]
      summary: Mint an API key for a tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, name]
              properties:
                tenant_id: { type: string, format: uuid }
                name:      { type: string, example: "booking-prod-key" }
                scopes:
                  type: array
                  items: { type: string }
                  example: ["instances:*","messages:*","chat:*","webhooks:*"]
      responses:
        "201":
          description: Created. The `api_key` is shown only once.
          content:
            application/json:
              example:
                id: "0193...-uuid"
                tenant_id: "0193...-uuid"
                name: "booking-prod-key"
                api_key: "zwa_live_a1b2c3d4e5f6_...40hex..."
                warning: "This key will NOT be shown again."
  /v1/admin/api-keys/{id}:
    delete:
      tags: [Admin]
      security: [{ AdminToken: [] }]
      summary: Revoke API key
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses: { "204": { description: Revoked } }

  # --------------------------------------------------------------- Instances
...